+++ conformity/resource_azure_account.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 )
 
 func resourceAzureAccount() *schema.Resource {
@@ -38,6 +39,58 @@ func resourceAzureAccount() *schema.Resource {
 					Type: schema.TypeString,
 				},
 			},
+			"settings": {
+				Type:     schema.TypeSet,
+				Optional: true,
+				MinItems: 0,
+				MaxItems: 1,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"bot": BotSettingsSchema(),
+						"rule": {
+							Type:     schema.TypeList,
+							Optional: true,
+							Elem: &schema.Resource{
+								Schema: map[string]*schema.Schema{
+									"note": {
+										Type:     schema.TypeString,
+										Required: true,
+									},
+									"rule_id": {
+										Type:     schema.TypeString,
+										Required: true,
+									},
+									"settings": {
+										Type:     schema.TypeSet,
+										Required: true,
+										MaxItems: 1,
+										Elem: &schema.Resource{
+											Schema: map[string]*schema.Schema{
+												"enabled": {
+													Type:     schema.TypeBool,
+													Optional: true,
+													Default:  true,
+												},
+												"rule_exists": {
+													Type:     schema.TypeBool,
+													Optional: true,
+												},
+												"exceptions":     ExceptionsSchema(),
+												"extra_settings": ExtraSettingSchema(),
+												"risk_level": {
+													Type:         schema.TypeString,
+													Required:     true,
+													ValidateFunc: validation.StringInSlice([]string{"LOW", "MEDIUM", "HIGH", "VERY_HIGH", "EXTREME"}, false),
+												},
+											},
+										},
+									},
+								},
+							},
+						},
+					},
+				},
+			},
 		},
 		Importer: &schema.ResourceImporter{
 			StateContext: schema.ImportStatePassthroughContext,
@@ -61,6 +114,11 @@ func resourceAzureAccountCreate(ctx context.Context, d *schema.ResourceData, m i
 		return diag.FromErr(err)
 	}
 
+	err = updateAccountSettings("azure", accountId, d, client)
+	if err != nil {
+		return diag.FromErr(err)
+	}
+
 	d.SetId(accountId)
 	resourceAzureAccountRead(ctx, d, m)
 	return diags
@@ -78,6 +136,11 @@ func resourceAzureAccountRead(ctx context.Context, d *schema.ResourceData, m int
 		return diag.FromErr(err)
 	}
 
+	ruleSettings, err := client.GetAccountRuleSettings(accountId)
+	if err != nil {
+		return diag.FromErr(err)
+	}
+
 	if err := d.Set("name", accountDetails.Data.Attributes.Name); err != nil {
 		return diag.FromErr(err)
 	}
@@ -87,6 +150,11 @@ func resourceAzureAccountRead(ctx context.Context, d *schema.ResourceData, m int
 	if err := d.Set("subscription_id", accountDetails.Data.Attributes.CloudData.Azure.SubscriptionId); err != nil {
 		return diag.FromErr(err)
 	}
+	// rule setting for azure is not yet done
+	settings := flattenAccountSettings(accountDetails.Data.Attributes.Settings, ruleSettings.Data.Attributes.Settings.Rules)
+	if err := d.Set("settings", settings); err != nil {
+		return diag.FromErr(err)
+	}
 	return diags
 }
 
@@ -94,6 +162,7 @@ func resourceAzureAccountUpdate(ctx context.Context, d *schema.ResourceData, m i
 	client := m.(*cloudconformity.Client)
 	// Warning or errors can be collected in a slice type
 	var diags diag.Diagnostics
+	accountId := d.Id()
 	if d.HasChange("name") || d.HasChange("environment") || d.HasChange("tags") {
 		payload := cloudconformity.AccountPayload{}
 		payload.Data.Attributes.Name = d.Get("name").(string)
@@ -104,12 +173,19 @@ func resourceAzureAccountUpdate(ctx context.Context, d *schema.ResourceData, m i
 			payload.Data.Attributes.Tags = append(payload.Data.Attributes.Tags, tag.(string))
 		}
 
-		accountId := d.Id()
 		_, err := client.UpdateAccount(accountId, payload)
 		if err != nil {
 			return diag.FromErr(err)
 		}
 	}
+
+	if d.HasChange("settings") {
+		err := updateAccountSettings("azure", accountId, d, client)
+		if err != nil {
+			return diag.FromErr(err)
+		}
+	}
+
 	// by activating the Active Directory Id need to add active_directory_id to this statement
 	// for now create function is not yet working so we need to remove active_directory_id for the if statement
 	if d.HasChange("subscription_id") {