+++ example/aws/provider.tf
@@ -1,7 +1,7 @@
 terraform {
   required_providers {
     conformity = {
-      version = "0.1.0"
+      version = "0.3.0"
       source  = "trendmicro.com/cloudone/conformity"
     }
       aws = {