+++ example/report_config/main/variable.tf
@@ -6,3 +6,13 @@ variable "region"{
     type    = string
     default = ""
 }
+// if you want to create organisation-level report config, leave account_id and group_id commented.
+variable "account_id" {
+    type    = string 
+    default = ""
+}
+// if you want to create group-level report config, uncomment group_id value and provide account ID on the terraform.tfvars
+// variable "group_id" {
+//     type    = string 
+//     default = ""
+// }